refactor: Admin, User의 Get Classified Major Service 분리 적용
- 인터페이스 분리 법칙 준수
