+++ src/main/java/com/server/computer_science/question/major_question/admin/controller/AdminMajorQuestionViewController.java
@@ -1,10 +1,8 @@
 package com.server.computer_science.question.major_question.admin.controller;
 
 
-import com.server.computer_science.question.major_question.admin.service.implement.BasicAdminMajorQuestionClassifiedGetService;
-import com.server.computer_science.question.major_question.common.service.MajorQuestionClassifiedGetService;
+import com.server.computer_science.question.major_question.admin.service.AdminMajorQuestionClassifiedGetService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -19,12 +17,12 @@ public class AdminMajorQuestionViewController {
     @Value("${resource.base-url}")
     private String resourceBaseUrl;
     private final String baseUrl = "baseUrl";
-    @Qualifier("basicAdminMajorQuestionClassifiedGetService")
-    private final MajorQuestionClassifiedGetService majorQuestionClassifiedGetService;
+
+    private final AdminMajorQuestionClassifiedGetService adminMajorQuestionClassifiedGetService;
 
     @GetMapping("/question/update")
     public String updateQuestionPage(Model model){
-        model.addAttribute("classifiedQuestions", majorQuestionClassifiedGetService.getClassifiedAllMajorQuestions());
+        model.addAttribute("classifiedQuestions", adminMajorQuestionClassifiedGetService.getClassifiedAllMajorQuestions());
         model.addAttribute(baseUrl, resourceBaseUrl);
         return "major-question-update";
     }
