+++ src/main/java/com/server/computer_science/question/major_question/admin/service/implement/BasicAdminMajorQuestionClassifiedGetService.java
@@ -1,11 +1,12 @@
 package com.server.computer_science.question.major_question.admin.service.implement;
 
 
+import com.server.computer_science.question.major_question.admin.service.AdminMajorQuestionClassifiedGetService;
 import com.server.computer_science.question.major_question.common.domain.MajorMultipleChoiceQuestion;
 import com.server.computer_science.question.major_question.common.service.QuestionClassifyByCategoryService;
 import com.server.computer_science.question.major_question.user.dto.request.RequestGetQuestionByCategoryAndLevelDto;
 import com.server.computer_science.question.common.dto.response.ResponseClassifiedMultipleQuestionDto;
-import com.server.computer_science.question.major_question.common.service.MajorQuestionClassifiedGetService;
+import com.server.computer_science.question.major_question.user.service.MajorQuestionClassifiedGetService;
 import com.server.computer_science.question.major_question.common.service.implement.MajorMultipleChoiceQuestionDBService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,20 +14,11 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Service("basicAdminMajorQuestionClassifiedGetService")
+@Service
 @RequiredArgsConstructor
-public class BasicAdminMajorQuestionClassifiedGetService implements MajorQuestionClassifiedGetService {
+public class BasicAdminMajorQuestionClassifiedGetService implements AdminMajorQuestionClassifiedGetService {
     private final MajorMultipleChoiceQuestionDBService majorMultipleChoiceQuestionDBService;
     private final QuestionClassifyByCategoryService questionClassifyByCategoryService;
-
-    @Override
-    public List<ResponseClassifiedMultipleQuestionDto> getClassifiedMajorMultipleChoiceQuestions(RequestGetQuestionByCategoryAndLevelDto requestGetQuestionByCategoryAndLevelDto) {
-        List<MajorMultipleChoiceQuestion> majorMultipleChoiceQuestions = majorMultipleChoiceQuestionDBService.getFetchChoicesByCategoriesAndLevels(requestGetQuestionByCategoryAndLevelDto.getQuestionCategories(), requestGetQuestionByCategoryAndLevelDto.getQuestionLevels());
-        return questionClassifyByCategoryService.classifyQuestionByCategoryOrdered(majorMultipleChoiceQuestions)
-                .entrySet().stream()
-                .map(entry-> ResponseClassifiedMultipleQuestionDto.forAdmin(entry.getKey(),entry.getValue()))
-                .collect(Collectors.toList());
-    }
 
     /**
      * 관리자가 조회시 Approve됐는지 기준으로 정렬되며(false인 것부터),
@@ -42,12 +34,5 @@ public class BasicAdminMajorQuestionClassifiedGetService implements MajorQuestio
                 .collect(Collectors.toList());
     }
 
-    @Override
-    public List<ResponseClassifiedMultipleQuestionDto> getClassifiedShortAnsweredMajorQuestions(RequestGetQuestionByCategoryAndLevelDto requestGetQuestionByCategoryAndLevelDto) {
-        List<MajorMultipleChoiceQuestion> majorMultipleChoiceQuestions = majorMultipleChoiceQuestionDBService.findAllFetchChoicesShortAnswered();
-        return questionClassifyByCategoryService.classifyQuestionByCategoryOrdered(majorMultipleChoiceQuestions)
-                .entrySet().stream()
-                .map(entry-> ResponseClassifiedMultipleQuestionDto.forAdmin(entry.getKey(),entry.getValue()))
-                .collect(Collectors.toList());
-    }
+
 }
