+++ src/test/java/com/server/computer_science/question/major_question/user/controller/MajorQuestionGetViewControllerTest.java
@@ -1,0 +1,15 @@
+package com.server.computer_science.question.major_question.user.controller;
+
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@WebMvcTest(MajorQuestionGetViewController.class)
+class MajorQuestionGetViewControllerTest {
+//    @MockBean
+//    @Qualifier("basicMajorQuestionClassifiedGetService")
+
+
+}
